djaein/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
Django==2.0.6
django-celery-beat==1.1.1
django-nose==1.4.5
djangorestframework==3.8.2
kombu==4.2.1
nose==1.3.7
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.4
vine==1.1.4
